Time: 97 ms (54.11%) | Memory: 91.3 MB (90.45%) - LeetSync

+++ LeetCode/3332-minimum-operations-to-exceed-threshold-value-ii/minimum-operations-to-exceed-threshold-value-ii.cpp
@@ -11,8 +11,8 @@ public:
         int count = 0;
 
         while(pq.top()<k){
-            long long x=pq.top(); pq.pop();
-            long long y=pq.top(); pq.pop();
+            long x=pq.top(); pq.pop();
+            long y=pq.top(); pq.pop();
             pq.push(2*x+y);
             count++;
         }
